♻️ refactor(book/header): move SelectedBookHeader to a separate subdirectory
- change import statements in frame.tsx, sentence.tsx, talk.tsx and sentence.tsx to reflect new file location
