+++ apps/tk-capture-book/app/(auth)/talk.tsx
@@ -3,7 +3,7 @@ import { View } from "react-native";
 import { useAtomValue } from "jotai";
 import { useTranslation } from "react-i18next";
 
-import { SelectedBookHeader } from "@/features/book/components/selected-book-header";
+import { SelectedBookHeader } from "@/features/book/components/selected-book/selected-book-header";
 import { BookTalk } from "@/features/book/components/talk/book-talk";
 import { BookTalkGuide } from "@/features/book/components/talk/book-talk-guide";
 import { selectedBookAtom } from "@/features/book/states/book";
